requirements.txt
aiofiles==23.1.0
aiosqlite==0.19.0
alembic==1.10.4
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
cffi==1.15.1
cryptography==40.0.2
databases==0.7.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.7.0
python-dateutil==2.8.2
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
SQLAlchemy-Utils==0.41.1
starlette==0.26.1
typing==3.7.4.3
typing_extensions==4.5.0
